Move cfPopHook to ClientState
